refactor: added static header data
