+++ pages/_app.tsx
@@ -1,17 +1,17 @@
 import type { AppProps } from 'next/app';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { Layout } from '../components/Layout';
 
 import '../styles/globals.css';
 
-import { FooterSection } from '../components/Footer';
-import { NavHeader } from '../components/NavHeader';
-
 function MyApp({ Component, pageProps }: AppProps) {
+  const client = new QueryClient();
   return (
-    <>
-      <NavHeader />
-      <Component {...pageProps} />;
-      <FooterSection />
-    </>
+    <QueryClientProvider client={client}>
+      <Layout>
+        <Component {...pageProps} />;
+      </Layout>
+    </QueryClientProvider>
   );
 }
